Pass InstanceInfo to the plug/unplug methods
Long term the VIF object instances will be created by Neutron,
serialized, and passed over to Nova. Only Nova has information
about the instance, so we can't pass the InstanceInfo object
in as an attribute on the VIFBase class when constructing it
in Neutron. Remove the 'instance_info' attribute from the
VIFBase class and instead pass it as a parameter to the
plug/unplug methods in the Plugin class.

Change-Id: I387a283d0e234c6a38118e0c7da754598855fbda