vif_plug_ovs: merge both plugins into one
There is no need to create separate plugins for each type of
VIF configuration - a single plugin is intended to be capable
of supporting multiple different VIF configs. On this basis,
the ovs & ovs_hybrid plugins can be merged to just create a
single 'ovs' plugin. This means that consumers of os-vif do
not need to remember multiple different plugin names - they
can always just use the 'ovs' plugin and feed it whichever
type of VIF config they desire.

Change-Id: I8ab37fea065a3d92585decf89bcc095c404e6ed2
Signed-off-by: Daniel P. Berrange <berrange@redhat.com>