os_vif: register objects before loading plugins
It is conceivable that some part of the plugin loading process
may wish to use the os_vif object model. We should thus ensure
we register objects, before we load any of the plugins. It is
also incorrect to re-register the plugins if an app asks os_vif
to reinitialize itself, as it'll lead to duplicate object
registrations. So we now only register objects the first
time initialize() is called.

Change-Id: Ib5730135636acf37e3f9ed67fe1d4533a2df3474