vif_plug_ovs: clarify that the plugin was not in fact renamed
An earlier change claimed to be renaming the ovs plugin:

  commit 65ae52baa712365eeb01d105004e86d83304da3c
  Author: Sean Mooney <sean.k.mooney@intel.com>
  Date:   Fri May 20 20:43:14 2016 +0100

    This change renames the ovs plugin

On re-examination this was not in fact the case. The actual
name of the plugins is specified in the setup.cfg config
file "[entry_points]" section, which recorded the name as
"ovs". The bug fixed in 65ae52baa712365eeb01d105004e86d83304da3c
was simply that the 'describe' method was reporting the wrong
name, which has no functional impact on usage of the plugin.
It merely meant that any app relying on the introspection
data to identify available plugins would be told the wrong
name.

Change-Id: Iaf7c2e38bce809afc4801a3af0971977461d66d3