os_vif: fix logging of exceptions during plug/unplug
The parameters to interpolate in the log message were being
passed directly, instead of in a dict, so they were never
actually merged into the message. While fixing this, we
can also avoid manually including the exception in the
message and leave it upto python logging by simply setting
exc_info=True.

Change-Id: Ib6adb52fb331ed148a6c250fd008e9c117f4cf68