Ensure the OVS bridge exists when plugging
When a vif will be plugged into an OVS bridge, ensure that the bridge
exists during the plugging operation. In most cases, the OVS bridge will
already exist at the time of port creation, so this will be a no-op.
However, the OVS agent implementation for vlan-aware VMs will require
the os-vif plugin to create the OVS bridge if it doesn't already exist,
because vifs that correspond to trunk ports won't be directly plugged
into br-int.

Change-Id: I3cb61a5e842177a26f38c85d7782ecced4ac07da