Don't create extraneous linux bridge/veth pair for VIFOpenVSwitch
VIFOpenVSwitch vifs will be booted by libvirt directly on the target
bridge. Don't create an extraneous linux bridge with a veth pair
connection to the target bridge.

VIFOpenVSwitch inherits from VIFBridge, so the following check causes
_plug_bridge() to be incorrectly executed for vifs of type
VIFOpenVSwitch:

    if isinstance(vif, objects.vif.VIFBridge):
        self._plug_bridge(vif, instance_info)

Fix this to check for instances of VIFOpenVSwitch first and do nothing
in that case.

Closes-Bug: #1595614
Change-Id: I447395cc50aa8fc21fa714b76f4cb1f5e79846ae