Add a reminder to remove Route.interface field
Nova never sets the Route.interface value to anything but
None which fails with an error:

"ValueError: Fieldinterface' cannot be None"

This looks like a carry-over from the nova.network.model.Route
class which has an interface field which is set to None by default
but that field is never set to anything else in Nova, neither
for nova-network or Neutron.

Furthermore, it looks like 'interface' is not something that's
in the Route data model in Neutron either.

We don't hit this in the gate because the subnets we're testing
with don't have host_routes set.

The ValueError was fixed in Nova by not setting the attribute:

1d57c1fd53e930b02c3ce0e9914f95ef68dd1f87

This change adds a TODO to remove it in version 2.0 of the Route object.

Change-Id: Ib25a79514fe4335f4df222c02fefc9da62fe04ce
Closes-Bug: #1612812