Update reno for stable/newton
In addition to adding the new newton page, this fixes the existing
release notes build to show notes for unreleased versions, which ensures
that no poorly formatted release note breaks the release notes build.

It also fixes the toctree directive at the top of the release notes.

In order to support automatically updating the release notes when we
create stable branches, we want the pages to be in a standard order.
This patch updates the order to be reverse chronological, so the most
recent notes appear at the top.

Change-Id: I14bd797d49ff00cfca1d3113e5b0b87d66d42d70
Signed-off-by: Doug Hellmann <doug@doughellmann.com>