os-vif: add vif_name to VIFVHostUser class
When dealing with vhostuser ports, plugins need to know the
name of such ports in order to perform plug/unplug.
The current implementation of vhostuser ovs plug computes
this vif name directly in the plugin while this information
could be directly stored in the object class.

This patch adds this port name (vif_name) to the VIFVHostUser
class to prepare future rework of this vhostuser ovs plug and to
enable different vhostuser plug mechanisms (including fast path)
that needs this vif_name as well.

Signed-off-by: Francesco Santoro <francesco.santoro@6wind.com>

Change-Id: I45f73918d51d5fa64a0c0edd557be1fd41db55f7