vif_plug_ovs: Skip setting MTU on Windows when plugging devices
Hyper-V with OVS does not support setting the MTU
of a Virtual interface externally to the VM.

This change disable the use of ip tool to set
the MTU on the guest interface which will always
fail on Windows.

On Windows based compute nodes, MTU advertisement
and configuration will be delegated to
the guest os via the DHCP mechanism.

Closes-Bug: #1654616

Change-Id: I16b33fedd9773c37a0368264aecb5581153ba404