ovs-hybrid: should permanently keep MAC entries
The linux bridge installed for the particular vif type ovs-hybrid
should be configured to persistently keep the MAC learned from the
RARP packets sent by QEMU when starting on destination node. That to
avoid any break of the datapath during a live-migration. That because
at some point of the live-migration source and destination can be on
a same L2 and we could have the destination bridge learning from source.

Change-Id: I50775df3629a079e6ba6f167ebfa499ffa5c7cac
Closes-Bug: 1715317
Related-Bug: 1414559