Move 'ips' field from Subnet object to VIF object
'ips' field in Subnet object [1] is a legacy field from Nova networks.
This is still used in the virt Firewall, but is not used in os-vif
project. Furthemore, the VIF object used by Nova to extract the port
IPs [2] is not provided by os-vif.

The reason to move this field from Subnet to VIF is because the IP list
should belong to the VIF object, which is linked with the port. The
Subnet object contains the CIDR of the subnetwork, but shouldn't store
the IPs given to an specific port.

[1] https://github.com/openstack/os-vif/blob/1.7.0/os_vif/objects/subnet.py#L30
[2] https://github.com/openstack/nova/blob/16.0.1/nova/virt/firewall.py#L180-L181

Change-Id: I99a4df96ee8e72915bdf9fc1327b9575dfa94586
Closes-Bug: #1720175