Fix VF-rep lookup routine to use parent PF number
The current logic in get_representor_port() matches the VF number with
the one in 'phys_port_name' file. If and when multiple PFs on the same
card are configured in switchdev mode, we could end up returning a wrong
representor device. Fix this by including the parent PF number also in
this logic, when 'pfXvfY' (or vfY@pfX) format is used in 'phys_port_name'
file.

This patch also provides corresponding changes in unit tests. The tests
have been updated to validate 'phys_port_name' in 'pfXvfY' format (in
addition to the simple format - vfX or X). The test case for 2 PFs
specifically utilizes the 'pfXvfY' format, since otherwise it won't
be possible to disambiguate VFs with the same VF-ID under 2 PFs on the
same card (switch).

Change-Id: I5727878b63026a19fafcf7e71dda3ba480432fbf
Closes-Bug: #1743494
Signed-off-by: Sriharsha Basavapatna <sriharsha.basavapatna@broadcom.com>
(cherry picked from commit 84688a22db22039f5566c1439c3983db23672a45)