switch documentation job to new PTI
This is a mechanically generated patch to switch the documentation
jobs to use the new PTI versions of the jobs as part of the
python3-first goal.

See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

Fix tox.ini docs environment to build html and push to correct output
paths so that openstack-tox-docs jobs works.

Change-Id: Ib957e03d6b16d5fdb945e2be02ed75f0c7fa160b
Story: #2002586
Task: #24315