Convert hardcoded regexes to raw strings for py36
From the Python 3.6 release notes:

A backslash-character pair that is not a valid escape sequence now
generates a DeprecationWarning. Although this will eventually become a
SyntaxError, that will not be for several Python releases. (Contributed
by Emanuel Barry in bpo-27364.) [1]

This patch converts the remaining regexes in the ovs plugin to raw.

[1]: https://docs.python.org/3/whatsnew/3.6.html#deprecated-python-behavior

Change-Id: I640d636cb602de543e942ad7f2339e63a487a43b
Signed-off-by: Jan Gutter <jan.gutter@netronome.com>