Add "master" parameter to ip.set() API function
This parameter is present in the IpCommand abstract class and both
Windows and Linux implementations.

This change modifies the pyroute2 driver to lookup the
interface index of the bridge when setting the master paramater
on a port.

This change add a functional test to use the master paramater
to add an interface to a bridge and validate it works correctly.

Change-Id: I54ded32281c6a5b67b6ccb9f1bdb6e3fb6731049
Closes-Bug: #1817919