Resolve dependency issues
This is totally broken with the pip 20.3 resolver. Attempts to fix the
lower-constraints file manually didn't work out, so start specifying
only the direct dependencies here, relying on upper-constraints for the
given release to restrict everything else.

Two dependencies, python-subunit and testrepository, are removed as
these are no longer necessary. Another, hacking, is moved to 'tox.ini'
since we're not managing this via the u-c system and don't need this
here.

Change-Id: I4bdc0d422813ecee1566256ec237d34f200406ee
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>