update os-vif ci to account for devstack default changes
This change restores the os-vif-ovs-iptables job to deploying
with ml2/ovs follow the devstack default change to ovn.

This change replaces the os-vif-ovs job with an os-vif-ovn jobs.

This change also simplifies teh os-vif-ovs-base job by disabling
DVR since that does not change the behavior of os-vif.

This change reduces the job time out to 7800 and moves it
to os-vif-tempest-base.

Closes-Bug: #1933590
Change-Id: Icda02ec4b5be18c3646af7988f210baf4ceb0890