update job template to zed
This change updates the python job template to zed
this will disable py36 jobs and enable the default set
for zed. The repo presently used the yoga template
which now fails as a result of project dropping py36 support.

This change also increases the min python reqired to 3.8
to match the zed supported runtimes and adds a release note.

Change-Id: I50253be0c99df9b363da5f6cc0c8502536dbbb43