Do not add taps in trunk bridges to the dead vlan
When using neutron trunks, the tpi port is meant to be put in the dead
vlan, not the tap, therefore

1) we no longer add taps in trunk bridges to the dead vlan in order to
   avoid bug #2069543.

While working on this I realized a related problem:

In neutron we changed how the dead vlan is handled. Instead of simply
setting tag=4095 we set (beyond having a drop rule for tag 4095):

tag : 4095
trunks : [4095]
vlan_mode : trunk

We do this not with the intention of making the port a trunk, but with
the intention of dropping all traffic arriving on the port. For details
please see the following patches:

neutron $ git log --oneline --no-merges --grep='dead vlan' -2
0ddca28454 Make sure "dead vlan" ports cannot transmit packets
7aae31c9f9 Make the dead vlan actually dead

Therefore 2) this patch also starts handling ports in the dead vlan as
neutron does.

Change-Id: I4f80d94bc1f346fd0a7ad0b9d684792f1451cf24
Closes-Bug: #2069543
Related-Bug: #1734320
Related-Bug: #1930414
Related-Bug: #1959564
(cherry picked from commit 99814b280fbdf8169d291e003cf6dc5a80e08074)