Update gate jobs as per the 2025.1 cycle testing runtime
As per 2025.1 testing runtime[1], we need to test on Ubuntu
Noble (which will be taken care by depends-on tempest and devstack
patches to move base jobs to Noble) and at least single job to run on
Ubutnu Jammy (for smooth upgrade from previous releases).

This commit adds a new job to run on Jammy which can be removed
in the next cycle. Fixing some pep8 issue also.

Depends-On: https://review.opendev.org/c/openstack/tempest/+/932156

[1] https://governance.openstack.org/tc/reference/runtimes/2025.1.html

Change-Id: I03ebbf04da76ad8b9c6877a8c90f962dcd1e9541