OVS Trunk: Add bridge_name to external_ids
In case of trunk wiring os-vif is responsible for
creating and deleting the trunk bridge (see [1] and [2]).
It can happen that Neutron can't set the metadata on the
bridge by adding i.e. bridge_name field the external_ids.
In such case, when the trunk is deleted before all these
operations were finished in Neutron, the clean-up of patches
linking br-int and the trunk bridge will never happen.

This patch adds the bridge_name field to external_ids in case
it is a trunk_bridge to avoid such situation.

Code documentation patch on Neutron side: [3].

[1]: https://review.opendev.org/c/openstack/os-vif/+/841499
[2]: https://review.opendev.org/c/openstack/neutron/+/837780
[3]: https://review.opendev.org/c/openstack/neutron/+/949569

Related-Bug: #2087944
Change-Id: I6951c1bfb5a85b1f422a5f9a99b6ade17f897c6f